#367 move fast_grid_scan_plan to new module
