+++ src/artemis/experiment_plans/__init__.py
@@ -1,0 +1,1 @@
+"""This module contains the experimental plans which artemis can run."""
